+++ myapp_conveuntact/lib/main.dart
@@ -4,11 +4,10 @@ import 'package:myapp/pages/login/login_screen.dart';
 import 'package:get/get.dart';
 import 'package:get/get_navigation/src/root/get_material_app.dart';
 import 'package:firebase_core/firebase_core.dart';
-import 'package:myapp/pages/signup/admin_finish_signup.dart';
-import 'package:myapp/pages/signup/reset_password.dart';
-import 'package:myapp/pages/signup/select_army.dart';
-import 'package:myapp/pages/signup/add_detail.dart';
-import 'package:myapp/pages/signup/select_detail.dart';
+import 'package:myapp/pages/signup/admin/admin_reset_password.dart';
+import 'package:myapp/pages/signup/admin/select_army.dart';
+import 'package:myapp/pages/signup/admin/add_detail.dart';
+import 'package:myapp/pages/signup/admin/select_detail.dart';
 import 'package:myapp/pages/signup/signtype.dart';
 
 void main() async {
@@ -23,7 +22,7 @@ class MyApp extends StatelessWidget {
     return GetMaterialApp(
       debugShowCheckedModeBanner: false,
       // 라우트 설계 필요없음. GetX 사용할 예정
-      home: ResetPassword(),
+      home: (AdminResetPassword()),
     );
   }
 }
