+++ myapp_conveuntact/lib/pages/signup/user/user_finish_reset_password.dart
@@ -1,15 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:myapp/pages/login/login_screen.dart';
-import 'package:myapp/pages/signup/unitcode.dart';
 
-const unitcode = '211022';
-const primaryColor = Color(0xFFACBDF4);
-void main() => runApp(const AdminFinishSignup());
+const primaryColor = Color(0xFFF7CBD4);
+void main() => runApp(const UserFinishResetPassword());
 
 /// This is the main application widget.
-class AdminFinishSignup extends StatelessWidget {
-  const AdminFinishSignup({Key? key}) : super(key: key);
+class UserFinishResetPassword extends StatelessWidget {
+  const UserFinishResetPassword({Key? key}) : super(key: key);
 
   //static const String _title = 'Test TextButton';
 
@@ -17,7 +15,7 @@ class AdminFinishSignup extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       theme: ThemeData(
-        scaffoldBackgroundColor: Color(0xFFACBDF4),
+        scaffoldBackgroundColor: Color(0xFFF7CBD4),
         primaryColor: primaryColor,
       ),
       //title: _title,
@@ -55,32 +53,22 @@ class MyStatelessWidget extends StatelessWidget {
                   fontSize: 50,
                   color: Colors.white)),
           SizedBox(height: 120),
-          Text('환영합니다! \n관리자님',
+          Text('메일을 보냈어요',
               style: TextStyle(
                   //fontWeight: FontWeight.bold,
-                  fontSize: 70,
+                  fontSize: 50,
                   color: Colors.white)),
-          SizedBox(height: 20),
-          Text('회원가입이 성공적으로 완료되었습니다. \n부대코드는 ${unitcode} 입니다.',
+          SizedBox(height: 40),
+          Text('conveuntact@naver.com',
+              style: TextStyle(
+                  fontWeight: FontWeight.bold,
+                  fontSize: 20,
+                  color: Colors.blue)),
+          Text('메일의 링크를 클릭하여\n비밀번호를 초기화 하세요',
               style: TextStyle(
                   fontWeight: FontWeight.bold,
                   fontSize: 20,
                   color: Colors.black26)),
-          // TextButton(
-          //   style: TextButton.styleFrom(
-          //     textStyle: const TextStyle(fontSize: 20),
-          //   ),
-          //   onPressed: null,
-          //   child: const Text('사용자'),
-          // ),
-          //const SizedBox(height: 60),
-          // TextButton(
-          //   style: TextButton.styleFrom(
-          //     textStyle: const TextStyle(fontSize: 60),
-          //   ),
-          //   onPressed: () {},
-          //   child: const Text('관리자'),
-          // ),
           SizedBox(height: 130),
           ClipRRect(
             borderRadius: BorderRadius.circular(10),
@@ -108,7 +96,7 @@ class MyStatelessWidget extends StatelessWidget {
                   onPressed: () {
                     Get.to(Loginpage());
                   },
-                  child: const Text('시작하기'),
+                  child: const Text('   확인   '),
                 ),
               ],
             ),
