+++ myapp_conveuntact/lib/pages/signup/user/user_info.dart
@@ -1,25 +1,26 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:get/get.dart';
-import 'package:myapp/pages/signup/select_army.dart';
+import 'package:myapp/pages/signup/user/unitcode.dart';
+import 'user_check_unit.dart';
 
-void main() => runApp(AdminInfo());
-const primaryColor = Color(0xFFACBDF4);
+void main() => runApp(UserInfo());
+const primaryColor = Color(0xFFF7CBD4);
 
-class AdminInfo extends StatelessWidget {
+class UserInfo extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     //final appTitle = 'sign up page';
     return MaterialApp(
       theme: ThemeData(
-        scaffoldBackgroundColor: Color(0xFFACBDF4),
+        scaffoldBackgroundColor: Color(0xFFF7CBD4),
         primaryColor: primaryColor,
       ),
       //title: appTitle,
       home: Scaffold(
         appBar: AppBar(
           backgroundColor: primaryColor,
-          title: Text('관리자 개인정보 입력페이지'),
+          title: Text('사용자 개인정보 입력페이지'),
           automaticallyImplyLeading: true,
           leading: IconButton(
             icon: Icon(Icons.arrow_back_ios_new),
@@ -52,12 +53,12 @@ class MyCustomFormState extends State<MyCustomForm> {
     // Build a Form widget using the _formKey created above.
     return Column(
       children: [
-        SizedBox(height: 200),
+        SizedBox(height: 120),
         Form(
           key: _formKey,
           child: Container(
             margin: const EdgeInsets.all(30),
-            padding: EdgeInsets.all(30.0),
+            padding: EdgeInsets.all(10.0),
             decoration: BoxDecoration(
               color: Colors.white,
               border: Border.all(
@@ -69,23 +70,24 @@ class MyCustomFormState extends State<MyCustomForm> {
             child: Column(
               crossAxisAlignment: CrossAxisAlignment.center,
               children: <Widget>[
-                SizedBox(height: 70),
+                SizedBox(height: 40),
                 TextFormField(
                   decoration: const InputDecoration(
                     border: OutlineInputBorder(),
                     labelStyle: TextStyle(fontWeight: FontWeight.bold),
                     icon: const Icon(Icons.person),
-                    hintText: '부대번호를 입력해주세요.',
-                    labelText: '아이디(부대 번호)',
+                    hintText: '- 를 포함하여 입력해 주세요.',
+                    labelText: '아이디(군번)',
                   ),
                   validator: (value) {
                     if (value!.isEmpty) {
-                      return '부대번호를 입력해주세요';
+                      return '군번을 입력해주세요';
                     }
                     return null;
                   },
                 ),
                 const SizedBox(height: 10),
+
                 TextFormField(
                   decoration: const InputDecoration(
                     border: OutlineInputBorder(),
@@ -97,6 +99,38 @@ class MyCustomFormState extends State<MyCustomForm> {
                   validator: (value) {
                     if (value!.isEmpty) {
                       return '비밀번호를 입력해주세요';
+                    }
+                    return null;
+                  },
+                ),
+                const SizedBox(height: 10),
+                TextFormField(
+                  decoration: const InputDecoration(
+                    border: OutlineInputBorder(),
+                    labelStyle: TextStyle(fontWeight: FontWeight.bold),
+                    icon: const Icon(Icons.near_me),
+                    hintText: '이름을 입력해 주세요  ex) 홍길동',
+                    labelText: '이름',
+                  ),
+                  validator: (value) {
+                    if (value!.isEmpty) {
+                      return '이름을 입력해주세요';
+                    }
+                    return null;
+                  },
+                ),
+                const SizedBox(height: 10),
+                TextFormField(
+                  decoration: const InputDecoration(
+                    border: OutlineInputBorder(),
+                    labelStyle: TextStyle(fontWeight: FontWeight.bold),
+                    icon: const Icon(Icons.person_add_alt_rounded),
+                    hintText: '계급을 입력해 주세요  ex) 일병',
+                    labelText: '계급',
+                  ),
+                  validator: (value) {
+                    if (value!.isEmpty) {
+                      return '계급을 입력해주세요';
                     }
                     return null;
                   },
@@ -130,7 +164,7 @@ class MyCustomFormState extends State<MyCustomForm> {
                 //         }
                 //       },
                 //     )),
-                const SizedBox(height: 30),
+                const SizedBox(height: 60),
                 ClipRRect(
                   borderRadius: BorderRadius.circular(20),
                   child: Stack(
@@ -141,9 +175,9 @@ class MyCustomFormState extends State<MyCustomForm> {
                           decoration: const BoxDecoration(
                             gradient: LinearGradient(
                               colors: <Color>[
-                                Color(0xFFACBDF4),
-                                Color(0xFFACBDF4),
-                                Color(0xFFACBDF4),
+                                Color(0xFFF3D0D8),
+                                Color(0xFFF3D0D8),
+                                Color(0xFFF3D0D8),
                               ],
                             ),
                           ),
@@ -162,7 +196,7 @@ class MyCustomFormState extends State<MyCustomForm> {
                             // If the form is valid, display a Snackbar.
                             Scaffold.of(context)
                                 .showSnackBar(SnackBar(content: Text('완료.')));
-                            Get.to(SelectArmy());
+                            Get.to(UnitCode());
                           }
                         },
                         child: const Text('다음'),
